+++ CHANGES.md
@@ -1,3 +1,7 @@
+# 2025-09-24 (8.4.0)
+
+* Delete datasets when dataset.json schema is deleted
+
 # 2025-09-25 (8.3.6.post1)
 
 * Minor fix for error in scope filtering